pyproject.toml
[tool.poetry]
name = "python_api_development"
version = "0.1.0"
description = ""
authors = ["oswaldohernandez <oswaldohernandez@protonmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
pydantic = "^2.3.0"
psycopg = {extras = ["binary"], version = "^3.1.12"}
email-validator = "^2.0.0.post2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
types-passlib = "^1.7.7.13"
types-python-jose = "^3.3.4.8"
sqlalchemy = "^2.0.21"
psycopg2 = "^2.9.9"
pydantic-settings = "^2.0.3"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.9.1"
flake8 = "^6.1.0"
mypy = "^1.5.1"
pylint = "^2.17.6"
alembic = "^1.12.0"

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
